Fix ESP32-S3 compilation errors: snprintf type casting, build configuration, and sdkconfig generation
- Fixed snprintf ambiguous overload (libstdc++ vs embedded toolchain stdio.h) with explicit (size_t) casting
- Updated platform to espressif32@~6.8.0 with esp32-s3-devkitc-1 board configuration
- Corrected copy_sdkconfig.py to prevent header injection
- Applied build flags from working reference configuration (PSA Battery Emulator)
- Firmware successfully builds: 308KB flash usage, 19.6KB RAM usage

+++ src/main.cpp
@@ -504,7 +504,7 @@ void loop() {
       for (int i = 0; i < len; i++) {
         Serial.print(":");
 
-        snprintf(tmp, 3, "%02X", canMsgRcv.data[i]);
+        snprintf(tmp, (size_t)3, "%02X", canMsgRcv.data[i]);
 
         Serial.print(tmp);
       }
@@ -1625,7 +1625,7 @@ void loop() {
       for (int i = 0; i < len; i++) {
         Serial.print(":");
 
-        snprintf(tmp, 3, "%02X", canMsgRcv.data[i]);
+        snprintf(tmp, (size_t)3, "%02X", canMsgRcv.data[i]);
 
         Serial.print(tmp);
       }
